☀️ 20% – Run return-type recovery over the main image too
The rule is not overlay-specific: a callee's declared return type decides the
order of the pre-call argument moves whether the caller is an overlay row or a
main-image owner. Swept over all 755 main-image semantic owners with
`candidate-show` as the oracle, it closes `08093054` -- an owner an earlier
session had recorded as stuck at two halfwords after 74 mode combinations --
and moves 27 others closer, several by tens of halfwords.

`08093054` is the useful witness. It was parked as a scheduling residual and it
was a prototype: one declaration, one flip. Two halfwords of "the allocator
decided" turned out to be an interface we had guessed wrong.

267,272 -> 267,374 bytes. `make verify` green.

+++ semantic/08028808.c
@@ -1,7 +1,7 @@
 #include "types.h"
 
 s32 Func_080022ec(s32, s32);
-s32 Func_080162d4(s32, s32, s32, s32, s32);
+u8 Func_080162d4(s32, s32, s32, s32, s32);
 
 typedef struct State {
     u8 pad00[0x78];
